// Downloads/Khel-india-frontend-main/Khel-india-frontend-main/src/shared/typings/request.ts
export type QueryRequest<T> = {
  loading: boolean;
  refreshing?: boolean;
  error?: Error;
  data?: T;
};

export interface PaginatedQueryRequest<T, K> extends QueryRequest<Array<T>> {
  isFetchingMore: boolean;
  fetched?: boolean;
  paginationData?: K & { hasMore: boolean };
  hasMore?: boolean;
  refreshing?: boolean;
}

export type QueryRequestFamilyParam<T> = {
  id: string;
  data?: QueryRequest<T>;
};

export type PaginatedQueryRequestFamilyParam<T, K> = {
  id: string;
  data?: PaginatedQueryRequest<T, K>;
};

export type RequestFamilyParam<T> = {
  id: string;
  data?: QueryRequest<T>;
};